dataset-shifts/requirements.txt
click==7.1.2
cycler==0.10.0
dataclasses==0.6
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mysql-connector-python==8.0.22
numba==0.51.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.21
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.52.0
typing-extensions==3.7.4.3
umap-learn==0.4.6
Werkzeug==1.0.1
